+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
+## [0.7.1] - 2026-02-22
+
+### Fixes
+
+* Update library images that were not updated in release 0.7.0
+* Build script updates all library images
+
 ## [0.7.0] - 2026-02-16
 
 ### Fixes